backend/requirements.txt
# Web Framework & Server
fastapi
uvicorn[standard]

# Database (MySQL Async)
sqlalchemy>=2.0
aiomysql
cryptography

# Cache & Task Queue
redis>=4.2.0

# AI & NLP
openai>=1.0.0
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
pyahocorasick

# Utilities
python-dotenv
pydantic
pydantic-settings

# Auth
PyJWT
passlib[bcrypt]
email-validator